\name{abcq}
\docType{data}
\alias{abcq}
\title{American Business Cycle}
\description{
The 22 variables are quarterly data from  1919-1941 and 1947-1983 of the variables GNP72, CPRATE, CORPYIELD, M1, M2, BASE, CSTOCK, WRICE67, PRODUR72, NONRES72, IRES72, DBUSI72, CDUR72, CNDUR72, XPT72, MPT72, GOVPUR72,NCSPDE72, NCSBS72, NCSCON72, CCSPDE72 and CCSBS72. 
}
\usage{abcq}
\format{A matrix of size  240 x 22}
\source{http://data.nber.org/data/abc/} 
\keyword{datasets}
