\name{boston}
\docType{data}
\alias{boston}
\title{Boston data}
\description{
  This data set is part of the MASS package.  The 14 columns are:

crim per capita crime rate by town

zn proportion of residential land zoned for lots over 25.000 sq.ft.

indus proportion of non-residential business acres per town

chas Charles River dummy variable (=1 if tract bounds rive; 0
otherwise)

nox nitrogen oxides concentration (parts per 10 million)

rm average number of rooms per dwelling

age proportion of owner-occupied units built prior to 1940

dis weighted mean of distances to five Boston employment centres

rad index of accessibility to radial highways

tax full-value property-tax rate per $10,000

ptration pupil-teacher ration by town

black 100(Bk-0.63)^2 where Bk is the proportion of blacks by town

lstat lower status of the population (percent)

medv median value of owner occupies homes in $1000s.
}

\usage{boston}
\format{A 506 x 14 matrix.}
\source{R package MASS
https://cran.r-project.org/web/packages/available_packages_by_name.html}
\references{
  MASS Support Functions and Datasets for Venables and Ripley's MASS
}
\keyword{datasets}
