\name{colon}
\docType{data}
\alias{colon.y}
\title{Colon data}
\description{
  The vector "colon.y" consists of 62 persons, 40 with
colon cancer (=1) and 22 controls (=0).
}

\usage{colon.y}
\format{a vector of length 62}
\source{http://microarray.princeton. edu/oncology
http://stat.ethz.ch/~dettling/bagboost.html
}
\references{Alon, U., Barkai, N., Notterman, D., Gish, K., Ybarra, S., Mack, D., and
  Levine, A. Broad patterns of gene expression revealed by clustering analysis of
  tumor and normal colon tissues probed by oligonucleotide arrays.
PNAS,  (1999), 96(12).
Booting for tumor classification with gene expression
data. Dettling, M. and B\"uhlmann, P., Bioinformatics, (2003),19(9):1061--1069.
}
\keyword{datasets}