\name{dent}
\docType{data}
\alias{dent}
\title{Dental data}
\description{
This is a 8x3x5 ANOVA table with the dependent variable being the hardness of a dental gold filling. There were eight different gold alloys, three different methods of preparation and five different dentists who  prepared the filling .
}
\usage{dent}
\format{A matrix of size  120 x 4}
\source{Towards robust analysis of variance, Seheult, A.H. and Tukey, J.W. (2001)
}
\references{Towards robust analysis of variance. Seheult, A.H. and
Tukey, J.W. In Saleh, A.K.M.E., editor, Data Analysis from Statistical  Foundations: A Festschrift in Honor of the 75th Birthday of D.A.S. Fraser,  (2001), 217--244.
}
\keyword{datasets}