\name{dentx}
\docType{data}
\alias{dentx}
\title{Dental data, the 13 covariates, the dependent variable is denty}
\description{
These are the covariates in a 8x3x5 ANOVA. There are eight different gold alloys, three different methods of preparation and five different dentists who  prepared the filling .The covariates 1:7 are the gold used, the covariate 8:9 are the method of condensation, the covariates 10:13 are the dentists. 
}
\usage{dentx}
\format{A matrix of dimension 120x13}
\source{Towards robust analysis of variance, Seheult, A.H. and Tukey, J.W. (2001)
}
\references{Towards robust analysis of variance. Seheult, A.H. and
Tukey, J.W. In Saleh, A.K.M.E., editor, Data Analysis from Statistical  Foundations: A Festschrift in Honor of the 75th Birthday of D.A.S. Fraser,  (2001), 217--244.
}
\keyword{datasets}