% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdecode.R
\name{fdecode}
\alias{fdecode}
\title{Decodes the number of a subset selected by fasb.R to give the  covariates} 
\usage{
fdecode(ns, k)
}
\arguments{
\item{ns}{The number of the subset}

\item{k}{The number of covariates}
}
\value{
ind The list of covariates

set A binary vector giving the covariates
}
\description{
Decodes the number of a subset selected by fasb.R to give the  covariates
}
\examples{
a<- fdecode(19,8)
}
