% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1st.R
\name{fgenbsf}
\alias{fgenbsf}
\title{Generates basis functions on disjoint intervals}
\usage{
fgenbsf(n,k,lam)
}
\arguments{
\item{n}{Sample size}

\item{k}{Size of smallest interval}

\item{lam}{Proportionality factor for increasing size of intervals}
}
\value{
x the matrix of the intervals
}
\description{
Generation of basic functions on intervals
}
\examples{
a<-fgenbsf(100,4,1.2)
}
