% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgentrig.R
\name{fgentrig}
\alias{fgentrig}
\title{generation of sine and cosine functions}
\usage{
fgentrig(n,m)
}
\arguments{
\item{n}{Sample size}
\item{m}{m sine and m cosine functions}
}
\value{
x  The functions sin(pi*j*(1:n)/n) and cos(pi*j*(1:n)/n) for j=1,...,m.
}
\description{
Generates the sine and cosine functions of order j, j=1,...,m.
}
\examples{
trig100<-fgentrig(100,50)[[1]]
}

