% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgr2st.R
\name{fgr2st}
\alias{fgr2st}
\title{Calculates an independence graph using  repeated stepwise selection}
\usage{
fgr2st(x,p0=0.01,km=0,kmx=0,nu=1,nedge=10^6,inr=T,dr=F)
}
\arguments{
\item{x}{The variables}

\item{p0}{Cut-off P-value}

\item{km}{Maximum number selected variables for each node irrespective of cut-off P-value}

\item{kmx}{The maximum number of selected covariates for each node for given cut-off P-value}

\item{nu}{The order statistic of Gaussian covariates used for comparison.}

\item{nedge}{Maximum number of edges}

\item{inr}{Logical, if TRUE include an intercept}

\item{dr}{Logical, if TRUE (a,b) and (b,a), a not equal b,  are different edges}

}
\value{
ned Number of edges

edg List of edges
}
\description{
Calculates a dependency graph using  repeated Gaussian stepwise selection
}
\examples{
data(redwine)
a<-fgr2st(redwine[,1:11]) 
}
