% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgrall.R
\name{fgrall}
\alias{fgrall}
\title{Calculates a dependence graph using Gaussian all subset selection}
\usage{
fgrall(x,p0=0.01,kmx=0,mx=21,inr=T,xinr=F)
}
\arguments{
\item{x}{The matrix of covariates}

\item{p0}{Cut-off P-value}

\item{kmx}{Maximum number  included covariates for each node irrespective of cut-off P-value}

\item{mx}{The maximum number of covariates.}

\item{inr}{Logical, if TRUE include an intercept}

\item{xinr}{Logical, if TRUE intercept already included}

}
\value{
ned Number of edges

edg List of edges with Gaussian P-value and percentage of sum of squared residuals explained by edge
}
\description{
Calculates an independence graph using Gaussian stepwise selection
}
\examples{
data(boston)
a<-fgrall(redwine[,1:8])
}
