% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpsired.R
\name{fpsired}
\alias{fpsired}
\title{Calculates Hampel's redescending psi function}
\usage{
fpsired(x,cnr)
}
\arguments{
\item{x}{The point at which the psi function is evaluated}

\item{cnr}{The parameters of Hampel's redescending psi function}

}
\value{
rpsi The value of the function
}
\description{
Calculates Hampel's redescending psi function
}
\examples{
fpsired(1,c(1,3,5))
}