% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobreg.R
\name{frobreg}
\alias{frobreg}
\title{Robust regression using Huber's psi-function}
\usage{
frobreg(y, x, cn, cpp = 0, sig = 0, intercept = TRUE)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Covariates}

\item{cn}{Tuning parameter for Huber's psi-function}

\item{cpp}{Fisher consistency parameter}

\item{sig}{Scale}

\item{intercept}{Logical to include intercept}
}
\value{
beta Regression coefficients

res  Residuals

sig  Scale
}
\description{
Robust regression using Huber's psi-function
}
\examples{
data(gausscov)
bostrob<-frobreg(boston[,14],boston[,1:13],1)
bostrob[[1]]
plot(bostrob[[2]])
}
