% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsimords.R
\name{fsimords}
\alias{fsimords}
\title{Simulates the number of false positives for given dimensions (n,k) and given order statistic nu}
\usage{
fsimords(n, k, alpha, nu, kmax, nsim = 100)
}
\arguments{
\item{n}{The dimension of dependent variable}

\item{k}{The number of covariates}

\item{alpha}{Cut-off p-value}

\item{nu}{Order statistic}

\item{nsim}{Number of simulations}

\item{kmax}{Maximum number of false positives}
}
\value{
res Histogram of number of false positives.

mn Mean number of false positives.

ss Standard deviation of number of false positives
}
\description{
Simulates the number of false positives for given dimensions (n,k) and given order statistic nu
}
\examples{
fsimords(100,1000,0.05,3,6)
}
