\name{leukeia}
\alias{leukemia}
\docType{data}
\title{Leukemia data set}
\description{
Dataset of \eqn{n=72} persons indicating presence or absence of leukemia (variable 3572) and \eqn{q=3571} gene expressions of the 72 persons (variables 1 to 3571)
}
\usage{data(leukemia)}
\format{
  \describe{
    \item{y}{0-1 data of individuals with and without leukemia.}
    \item{x}{covariates of the level of 3571 genes.}
  }
}
\source{http://stat.ethz.ch/~dettling/bagboost.html}
\references{Boosting for tumor classification with gene expression
data. Dettling, M. and Buehlmann, P. Bioinformatics, 2003,19(9):1061--1069.
}
\keyword{datasets}
