\name{ly.original}
\docType{data}
\alias{ly.original}
\title{Leukemia data}
\description{
  The 72 persons, 25 with leukemia (=1) and 47 controls (=0).
}
\usage{ly.original}
\format{A column vector of length 72}
\source{http://stat.ethz.ch/~dettling/bagboost.html}
\references{Boosting for tumor classification with gene expression
data. Dettling, M. and Buehlmann, P. Bioinformatics, 2003,19(9):1061--1069.
}
\keyword{datasets}