\name{lx.original}
\docType{data}
\alias{lx.original}
\title{Leukemia data}
\description{
  The measurements of gene expression of 3571 genes.  
}
\usage{lx.original}
\format{A 72 x 3571 matrix}
\source{http://stat.ethz.ch/~dettling/bagboost.html}
\references{Boosting for tumor classification with gene expression
data. Dettling, M. and B\"uhlmann, P. Bioinformatics, 2003,19(9):1061--1069.
}
\keyword{datasets}
