\name{lymphoma}
\alias{lymphoma}
\docType{data}
\title{Lymphoma data set}
\description{
Dataset of \eqn{n=62} persons with one of three different forms of lymphoma and \eqn{q=4026} gene expressions of the 62 persons
}
\usage{data(leukemia)}
\format{
  \describe{
    \item{y}{0-1-2 data of individuals with lymphoma.}
    \item{x}{covariates of the level of 4062 genes.}
  }
}
\source{R Package ‘spls'}
\references{Sparse partial least squares classification for high dimensional data. Dongjun Chung and Sunduz Keles.  Stat. Appl. Genet. Mol. Biol., (2010;9)}
\keyword{datasets}
