\name{m15005m}
\docType{data}
\alias{m15005m}
\title{m15005m data}
\description{
  Angle and photon counts in thin film x-ray refraction. 
}
\usage{m15005m}
\format{A matrix of size  7001 x 2, first component angle, second photon count}
\source{The data were provided by Professor Dieter Mergel, Faculty of Physics, University of Duisburg-Essen, Essen, Germany}
\keyword{datasets}