\name{mel-temp}
\docType{data}
\alias{mel_temp}
\title{Melbourne minimum temperature}
\description{
  The daily minimum temperature in Melbourne for the years 1981-1990.
}
\usage{mel_temp}
\format{A vector  of length 3650}
\source{
https://www.kaggle.com/paulbrabban/daily-minimum-temperatures-in-melbourne
}
\keyword{datasets}