\name{prostate.x}
\docType{data}
\alias{prostate.x}
\title{Prostate cancer data}
\description{
  The measurements of gene expression of 6033 genes of 102 persons 
}
\usage{prostate.x}
\format{A 6033 x 102 matrix}
\source{https://stat.ethz.ch/~dettling/bagboost.html}
\references{Booting for tumor classification with gene expression
data. Dettling, M. and B\"uhlmann, P., Bioinformatics, (2003),19(9):1061--1069.
}
\keyword{datasets}
