\name{prostate.y}
\docType{data}
\alias{prostate.y}
\title{Prostate cancer data}
\description{
  A vector of length 102 indicating healthy (=0) and cancer patients
(=1)
}
\usage{prostate.y}
\format{A 0-1 vector of length 102}
\source{https://stat.ethz.ch/~dettling/bagboost.html}
\references{Booting for tumor classification with gene expression
data. Dettling, M. and B\"uhlmann, P., Bioinformatics, (2003),19(9):1061--1069.
}
\keyword{datasets}
