\name{redwine}
\docType{data}
\alias{redwine}
\title{Redwine data}
\description{
  The subjective quality of wine on an integer scale from 1-10 (variable 12) together with 11 physicochemical properties
}
\usage{redwine}
\format{A matrix of size  1599 x 12}
\source{https://archive.ics.uci.edu/ml/machine-learning-databases/wine-quality/
}
\references{ Modeling wine preferences by data mining from
physicochemical properties, Cortez, P., Cerdeira, A., Almeida, F.,
Matos, T., and Reis, J., Decision Support Systems, Elsevier, 2009,47(4):547--553.
}
\keyword{datasets}