\name{snspt}
\docType{data}
\alias{snspt}
\title{Sunspot data}
\description{
  The average number of sunspots each month from January 1749 to January 2020: variable 1 year; variable 2, month; variable 3 number of sunspots.
}
\usage{snspt}
\format{A matrix of size  3253 x 7}
\source{WDC-SILSO, Royal Observatory of Belgium, Brussels}
\keyword{datasets}