\name{stackloss}
\docType{data}
\alias{stackloss}
\title{Stack loss  data}
\description{
  Stack loss (column 4) together with Air.Flow, Water.Temp,
Acid.Conc. columns 1-3 respectively.
}
\usage{stackloss}
\format{A 21 x 4 matrix}
\source{R}
\references{Brownlee, K. A. (1960, 2nd ed. 1965) Statistical Theory and Methodology in Science and Engineering. New York: Wiley. pp. 491–500. 
}
\keyword{datasets}
