\name{vardata}
\docType{data}
\alias{vardata}
\title{USA economics data}
\description{ United States economic data taken from the FRED-MD macroeconomic database with the NAs removed.182 indices each of length 256}
\usage{vardata}
\format{A matrix of size  256 X 182}
\source{https://research.stlouisfed.org/econ/mccracken/fred-databases} 
\keyword{datasets}
