\name{abcql}
\docType{data}
\alias{abcql}
\title{American Business Cycle}
\description{
The 22 variables are quarterly data from  1919-1941 of the variables
GNP72, CPRATE, CORPYIELD, M1, M2, BASE, CSTOCK, WRICE67, PRODUR72,
NONRES72, IRES72, DBUSI72, CDUR72, CNDUR72, XPT72, MPT72, GOVPUR72,
NCSPDE72, NCSBS72, NCSCON72, CCSPDE72 and CCSBS72.  Each of these is given with 16 lags.  
}
\usage{abcql}
\format{A matrix of size  224 x 353}
\source{http://data.nber.org/data/abc/} 
\keyword{datasets}
