% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decodes the number of a subset selected by fmch to give the  covariates} 
\usage{
decode(ns, k)
}
\arguments{
\item{ns}{The number of the subset}

\item{k}{The number of covariates}
}
\value{
ind The list of covariates

set A binary vector giving the covariates
}
\description{
Decodes the number of a subset selected by fmch to give the  covariates
}
\examples{
a<- decode(100,10)
}
