% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decodes the number of a subset selected by flmmdch to give the  covariates} 
\usage{
decode(j, k)
}
\arguments{
\item{j}{The number of the subset}

\item{k}{The number of covariates}
}
\value{
set A binary vector giving the covariates
}
\description{
Decodes the number of a subset selected by flmmdch to give the  covariates
}
\examples{
a<-decode(19,8)
}
