\name{dentx}
\docType{data}
\alias{dentx}
\title{Dental data, the 20 covariates}
\description{
These are the covariates in a 8x3x5 ANOVA table together with 7 interaction terms. There were eight different gold alloys, three different methods of preparation and five different dentists who  prepared the filling .The covariates 1:7 are the gold used, the covariate 8:9 are the method of condensation, the covariates 10:13 are the dentists. The covariates 14:20 are interaction terms for observations 14, 90, 93,96, 103 119 and 120. have interaction terms. with. 
}
\usage{dentx}
\format{A matrix of dimension 120x20}
\source{Towards robust analysis of variance, Seheult, A.H. and Tukey, J.W. (2001)
}
\references{Towards robust analysis of variance. Seheult, A.H. and
Tukey, J.W. In Saleh, A.K.M.E., editor, Data Analysis from Statistical  Foundations: A Festschrift in Honor of the 75th Birthday of D.A.S. Fraser,  (2001), 217--244.
}
\keyword{datasets}