% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeninter.R
\name{fgeninter}
\alias{fgeninter}
\title{generation of interactions}
\usage{
fgeninter(x, ord, intercept = TRUE)
}
\arguments{
\item{x}{Covariates}

\item{ord}{Order of interactions}

\item{intercept}{Logical to include intercept}
}
\value{
xx  All interactions of order at most ord.
}
\description{
generation of interactions
}
\examples{
data(gausscov)
bostinter<-fgeninter(boston[,1:13],7)[[1]]
dim(bostinter)
}
