% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeninter.R
\name{fgeninter}
\alias{fgeninter}
\title{Generation of interactions}
\usage{
fgeninter(x,ord,inr=T,idv=0)
}
\arguments{
\item{x}{Covariates}

\item{ord}{Order of interactions}

\item{inr}{Logical to include intercept}

\item{idv}{List of 0-1 dummy covariates}
}
\value{
xx  All interactions of order at most ord.
}
\description{
Generates all interactions of degree at most ord excluding powers of 0-1 covariates
}
\examples{
data(boston)
bostinter<-fgeninter(boston[,1:13],2)[[1]]
}
