% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeninter.R
\name{fgeninter}
\alias{fgeninter}
\title{generation of interactions}
\usage{
fgeninter(x, ord, inr = T)
}
\arguments{
\item{x}{Covariates}

\item{ord}{Order of interactions}

\item{inr}{Logical to include intercept}
}
\value{
xx  All interactions of order at most ord.
}
\description{
Generates all interactions of degree at most ord
}
\examples{
data(boston)
bostinter<-fgeninter(boston[,1:13],7)[[1]]
}
