% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeninter.R
\name{fgeninter}
\alias{fgeninter}
\title{Generation of interactions}
\usage{
fgeninter(x,ord)
}
\arguments{
\item{x}{Covariates}

\item{ord}{Order of interactions}

}
\value{
xx  All interactions of order at most ord.

intx Decomposes a given interaction covariate of xx
}
\description{
Generates all interactions of degree at most ord 
}
\examples{
data(boston)
bostint<-fgeninter(boston[,1:13],2)
}
