% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgentrig.R
\name{fgentrig}
\alias{fgentrig}
\title{Generation of sine and cosine functions}
\usage{
fgentrig(n,m)
}
\arguments{
\item{n}{Sample size}
\item{m}{Maximum order of sine and cosine functions}
}
\value{
x  The functions sin(pi*j*(1:n)/n) (odd) and cos(pi*j*(1:n)/n) (even) for j=1,...,m.
}
\description{
Generates sin(pi*j*(1:n)/n) (odd) and cos(pi*j*(1:n)/n) (even) for j=1,...,m for a given sample size n.
}
\examples{
trig<-fgentrig(36,36)
}

