% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselect.R
\name{fselect}
\alias{fselect}
\title{Selects the subsets specified by fasb.R and frasb.R.}
\usage{
fselect(nv, k)
}
\arguments{
\item{nv}{The subsets specified by fasb.R or frasb.R}

\item{k}{The variables}
}
\value{
ind The selected subsets.
}
\description{
All subsets which are a subset of a specified  subset are removed. The remaining subsets are ordered by the sum of squares of the residuals (fasb.R) or the scale (frasb.R)
}
\examples{
b<-fasb(redwine[,12],redwine[,1:5 ])[[1]]
a<-fselect(b,11)[[1]]
b[a,]
}
