% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselect.R
\name{fselect}
\alias{fselect}
\title{Selects the subsets specified by flmmdch. It is called by flmmdch}
\usage{
fselect(nv, k)
}
\arguments{
\item{nv}{The subsets specified by flmmdch}

\item{k}{The variables}
}
\value{
ind The selected subsets.
}
\description{
All subsets which are a subset of a specified  subset are removed. The remaining subsets are ordered by the sum of squares of the residuals
}
\examples{
data(gausscov)
nv<-flmmdch(redwine[,12],redwine[,1:11])[[1]]
fselect(nv,13)
}
