% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselect.R
\name{fselect}
\alias{fselect}
\title{Selects the subsets specified by fmch.}
\usage{
fselect(nv, k)
}
\arguments{
\item{nv}{The subsets specified by fmch}

\item{k}{The variables}
}
\value{
ind The selected subsets.
}
\description{
All subsets which are a subset of a specified  subset are removed. The remaining subsets are ordered by the sum of squares of the residuals
}
\examples{
nv<-c(650,1962,160,1033,394,1730,577,1839,334)
nv<-matrix(nv,ncol=3)
a<-fselect(nv,11)
}
