\name{leukemia}
\docType{data}
\alias{leukemia}
\title{Leukemia data}
\description{
Dataset of \eqn{n=72} persons indicating presence or absence of leukemia and \eqn{q=3571} gene expressions of the 72 persons}
\usage{data(leukemia)}
\format{
	\describe{
		item{lekemia[[1]]}{0-1 vector of length giving presence or absence of leukemia}
		item{leukemia[[2]]}{72x3571 matrix giving the gene exprassions of the 72 persons}
	}
}
\source{http://stat.ethz.ch/~dettling/bagboost.html}
\references{Boosting for tumor classification with gene expression
data. Dettling, M. and Buehlmann, P. Bioinformatics, 2003,19(9):1061--1069.
}
\keyword{datasets}
