\name{leukemia}
\docType{data}
\alias{leukemia}
\title{Leukemia data}
\description{
Dataset of \eqn{n=72} persons indicating presence or absence of leukemia (variable 3572) and \eqn{q=3571} gene expressions of the 72 persons (variables 1 to 3571)
}
\usage{data(leukemia)}
\format{A matrix of size 72 x 3572}
\source{http://stat.ethz.ch/~dettling/bagboost.html}
\references{Boosting for tumor classification with gene expression
data. Dettling, M. and Buehlmann, P. Bioinformatics, 2003,19(9):1061--1069.
}
\keyword{datasets}
